# docs/openapi.yaml
openapi: '3.0.3'
info:
  title: Microservices API
  version: '1.0.0'
  description: |
    API Gateway for the microservices monorepo.
    The Gateway is the only publicly exposed HTTP endpoint.
    All inter-service communication is via NATS JetStream.
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserProfile:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
          format: uuid
          example: '550e8400-e29b-41d4-a716-446655440000'
        name:
          type: string
          example: 'Alice Smith'
        email:
          type: string
          format: email
          example: 'alice@example.com'
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
...
    ErrorResponse:
      type: object
      required: [statusCode, error, code, message]
      properties:
        statusCode:
          type: integer
          example: 400
        error:
          type: string
          example: 'Bad Request'
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Request validation failed'
        requestId:
          type: string
          format: uuid
        details:
          type: array
          items:
            type: object
paths:
...
  /api/v1/users/me:
    get:
      summary: Get the authenticated user's profile
      description: |
        Returns the profile of the currently authenticated user.
        
        Authentication flow:
        1. Client sends a valid JWT in the Authorization header.
        2. Gateway verifies the JWT signature and expiry.
        3. Gateway extracts the userId from the JWT `sub` claim.
        4. Gateway forwards userId to User Service via X-User-Id header (internal).
        5. User Service fetches and returns the profile.
        
        The client-provided userId is NEVER trusted — only the Gateway-derived value is used.
      operationId: getMyProfile
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Missing or invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                statusCode: 401
                error: 'UnauthorizedError'
                code: 'UNAUTHORIZED'
                message: 'Missing or malformed Authorization header'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: User Service unavailable
...